+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This page only documents changes to the library itself, **not** the icons. Please refer to the [mdi history](https://materialdesignicons.com/history) for that.
 
+## Unreleased [![Material Design Icons version](https://img.shields.io/badge/mdi-v2.3.54-blue.svg?style=flat-square)](https://materialdesignicons.com)
+
+### New features
+
+- Support for Preact via `mdi-preact` package.
+
 ## 3.2.0 [![Material Design Icons version](https://img.shields.io/badge/mdi-v2.3.54-blue.svg?style=flat-square)](https://materialdesignicons.com)
 
 _No changes_